Fixed error with VS2010 compilation. vsprintf was renamed to avoid name to clash with standard library. Fixed multiple warning related to x86/x64. Fixed range-based generics introduced in r9818 compilation under VS2010.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@9824 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_chart.cpp
@@ -1536,10 +1536,10 @@ const char* chart_t::timeline( std::string& s,
                                double avg,
                                const char* color )
 {
-  static const int max_points = 600;
+  static const size_t max_points = 600;
   static const double timeline_range  = 60.0;
 
-  int max_buckets = timeline_data.size();
+  size_t max_buckets = timeline_data.size();
   int increment = ( ( max_buckets > max_points ) ?
                      ( ( int ) floor( ( double ) max_buckets / max_points ) + 1 ) :
                      1 );
@@ -1571,7 +1571,7 @@ const char* chart_t::timeline( std::string& s,
   s += "chg=100,20";
   s += "&amp;";
   s += "chd=s:";
-  for ( int i=0; i < max_buckets; i += increment )
+  for ( size_t i=0; i < max_buckets; i += increment )
   {
     s += simple_encoding( ( int ) ( timeline_data[ i ] * timeline_adjust ) );
   }
